dom/security/test/csp/test_CSP.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Content Security Policy Connections</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<iframe style="width:200px;height:200px;" id='cspframe'></iframe>
<script class="testbody" type="text/javascript">

// These are test results: -1 means it hasn't run,
// true/false is the pass/fail result.
window.tests = {
  img_good: -1,
  img_bad: -1,
  style_good: -1,
  style_bad: -1,
  frame_good: -1,
  frame_bad: -1,
  script_good: -1,
  script_bad: -1,
  xhr_good: -1,
  xhr_bad: -1,
  fetch_good: -1,
  fetch_bad: -1,
  beacon_good: -1,
  beacon_bad: -1,
  worker_xhr_same_bad: -1,
  worker_xhr_cross_good: -1,
  worker_fetch_same_bad: -1,
  worker_fetch_cross_good: -1,
  worker_script_same_good: -1,
  worker_script_cross_bad: -1,
  worker_inherited_xhr_good: -1,
  worker_inherited_xhr_bad: -1,
  worker_inherited_fetch_good: -1,
  worker_inherited_fetch_bad: -1,
  worker_inherited_script_good: -1,
  worker_inherited_script_bad: -1,
  worker_child_xhr_same_bad: -1,
  worker_child_xhr_cross_bad: -1,
  worker_child_script_same_bad: -1,
  worker_child_script_cross_bad: -1,
  worker_child_inherited_parent_xhr_bad: -1,
  worker_child_inherited_parent_xhr_good: -1,
  worker_child_inherited_parent_script_good: -1,
  worker_child_inherited_parent_script_bad: -1,
  worker_child_inherited_document_xhr_good: -1,
  worker_child_inherited_document_xhr_bad: -1,
  worker_child_inherited_document_script_good: -1,
  worker_child_inherited_document_script_bad: -1,
  media_good: -1,
  media_bad: -1,
  font_good: -1,
  font_bad: -1,
  object_good: -1,
  object_bad: -1,
};

// This is used to watch the blocked data bounce off CSP and allowed data
// get sent out to the wire.
function examiner() {
  SpecialPowers.addObserver(this, "csp-on-violate-policy", false);
  SpecialPowers.addObserver(this, "specialpowers-http-notify-request", false);
}
examiner.prototype  = {
  observe: function(subject, topic, data) {
    var testpat = new RegExp("testid=([a-z0-9_]+)");

    //_good things better be allowed!
    //_bad things better be stopped!

    // This is a special observer topic that is proxied from
    // http-on-modify-request in the parent process to inform us when a URI is
    // loaded
    if (topic === "specialpowers-http-notify-request") {
      var uri = data;
      if (!testpat.test(uri)) return;
      var testid = testpat.exec(uri)[1];

      window.testResult(testid,
                        /_good/.test(testid),
                        uri + " allowed by csp");
    }

    if (topic === "csp-on-violate-policy") {
      // these were blocked... record that they were blocked
      var asciiSpec = SpecialPowers.getPrivilegedProps(SpecialPowers.do_QueryInterface(subject, "nsIURI"), "asciiSpec");
      if (!testpat.test(asciiSpec)) return;
      var testid = testpat.exec(asciiSpec)[1];
      window.testResult(testid,
                        /_bad/.test(testid),
                        asciiSpec + " blocked by \"" + data + "\"");
    }
  },

  // must eventually call this to remove the listener,
  // or mochitests might get borked.
  remove: function() {
    SpecialPowers.removeObserver(this, "csp-on-violate-policy");
    SpecialPowers.removeObserver(this, "specialpowers-http-notify-request");
  }
}

window.examiner = new examiner();

window.testResult = function(testname, result, msg) {
  // test already complete.... forget it... remember the first result.
  if (window.tests[testname] != -1)
    return;

  ok(testname in window.tests, "It's a real test");
  window.tests[testname] = result;
  is(result, true, testname + ' test: ' + msg);

  // if any test is incomplete, keep waiting
  for (var v in window.tests)
    if(tests[v] == -1)
      return;

  // ... otherwise, finish
  window.examiner.remove();
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPrefEnv(
  {'set':[// This defaults to 0 ("preload none") on mobile (B2G/Android), which
          // blocks loading the resource until the user interacts with a
          // corresponding widget, which breaks the media_* tests. We set it
          // back to the default used by desktop Firefox to get consistent
          // behavior.
          ["media.preload.default", 2]]},
    function() {
      // save this for last so that our listeners are registered.
      // ... this loads the testbed of good and bad requests.
      document.getElementById('cspframe').src = 'file_main.html';
    });
</script>
</pre>
</body>
</html>
